# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
paths:
...
  /api/v1/auth/verify-user-email:
    post:
      summary: 'Verify user email'
      operationId: verifyUserEmail
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'Must be a valid email address. Must not be greater than 255 characters.'
                  example: donnell.deckow@example.net
                token:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: wxdfzqdkjypiczzfuve
              required:
                - email
                - token
      security: []
...